+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added refreshThread method to Context
+
 ## v0.4.0 (2024-05-22)
 * Improved delete methods on Context
 * Fixed limit parameter placement in Context