+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.4.1 (2024-07-05)
+* Added refreshThread method to Context
+
 ## v0.4.0 (2024-05-22)
 * Improved delete methods on Context
 * Fixed limit parameter placement in Context